Changed title formatting & file location

+++ pageController.js
@@ -11,12 +11,15 @@ async function scrapeAll(browserInstance) {
     await browser.close();
     let spinner = ora("saving files").start();
     for (let entry of scrapedData) {
-      let showTitle = entry["showTitle"].split(" "),
-        shortTitle = "";
-      for (let i = 3, k = showTitle.length - 3; i < k; i++) {
-        shortTitle += showTitle[i] + " ";
-      }
-      shortTitle = shortTitle.trim();
+      /*let showTitle = entry["showTitle"].split(" "),
+        fileName = "";
+       for (let i = 3, k = showTitle.length - 3; i < k; i++) {
+        fileName += showTitle[i] + " ";
+      } */
+      let fileName =
+        entry.releaseDate.split(" ").slice(2).reverse().join("-") +
+        " – " +
+        entry.showTitle;
       let showResult = "";
       for (let i = 0, k = entry["artists"].length; i < k; i++) {
         const artistAndTack = `${i + 1}. ${entry["artists"][i]} – ${
@@ -26,13 +29,17 @@ async function scrapeAll(browserInstance) {
       }
       // path where I want the file on my computer
       const pathToShow = path.join(
-        "../../../Music/Iggy Confidential",
-        shortTitle
+        "../../../Music/Iggy Confidential/tracklists",
+        fileName
       );
       // save each show with proper title and the collected result as txt file
-      fs.writeFileSync(`${pathToShow}.txt`, showResult, "utf-8", (err) => {
-        if (err) return console.log(err);
-      });
+      // check if file exists already
+      let fileExists = fs.existsSync(`${pathToShow}.txt`);
+      if (!fileExists) {
+        fs.writeFileSync(`${pathToShow}.txt`, showResult, "utf-8", (err) => {
+          if (err) return console.log(err);
+        });
+      }
     }
     spinner.succeed().stop();
   } catch (error) {
